fix(helpers): Make work when using verifying doubles.
Errors were being thrown if the spec/rails helper was forcing the verification
of doubles (which is the default behaviour for RSpec 3). It was caused by having

config.mock_with :rspec do |mocks|
  mocks.verify_partial_doubles = true
end

somewhere in the config.

To resolve the way the stubbed flag is stored has been changed. It now requires
an arbitrary environment key, but it shouldn't conflict with existing keys.

Fixes #1